Move retrieve_date_time_info from C0 to B0
